Adaptive bfoa
TODO
Check how to do the adaptation (end of the loop), especially the e
variable
Check if correct data is used (compare for example the P and J variables
of bfoa. I removed some dimensions because there are less loops. But
need to check if it works like we expect)